requirements.txt
# Dependencies for the Tool Memory project
pymongo==4.6.1
letta==0.1.0
tavily-python==0.2.8
voyageai==0.1.6
python-dotenv==1.0.0
langchain==0.0.335
pydantic==2.5.2
rich==13.7.0
